Factor out v1 tests to common file
so we can do the same set of tests for the v1/v2 common apis
