+++ syditest/test_v1.py
@@ -1,0 +1,33 @@
+#!/usr/bin/env python
+
+# -*- coding: utf-8 -*-
+
+# Copyright 2019 The Matrix.org Foundation C.I.C.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import unittest
+
+from .is_api import IsApi
+from .base_api_test import BaseApiTest
+
+
+class V1Test(BaseApiTest, unittest.TestCase):
+    API_VERSION = 'v1'
+    #def __init__(self, *args, **kwargs):
+    #    super(V1Test, self).__init__(*args, **kwargs)
+    #    self.api = IsApi('v1')
+
+if __name__ == '__main__':
+    log.startLogging(sys.stdout)
+    unittest.main()
